make next "reexport" from server, instead of server exporting edge which exports next
