+++ packages/uploadthing/src/server.ts
@@ -1,17 +1,78 @@
+import { getStatusCodeFromError, UploadThingError } from "@uploadthing/shared";
 import type { Json } from "@uploadthing/shared";
 
-import { buildPermissionsInfoHandler } from "./internal/handler";
+import { UPLOADTHING_VERSION } from "./constants";
+import { defaultErrorFormatter } from "./internal/error-formatter";
+import type { RouterWithConfig } from "./internal/handler";
+import {
+  buildPermissionsInfoHandler,
+  buildRequestHandler,
+} from "./internal/handler";
 import type { FileRouter } from "./internal/types";
 import type { CreateBuilderOptions } from "./internal/upload-builder";
 import { createBuilder } from "./internal/upload-builder";
 
 export * from "./internal/types";
 export * as utapi from "./sdk";
-export { createServerHandler } from "./internal/edge";
 
 export const createUploadthing = <TErrorShape extends Json>(
   opts?: CreateBuilderOptions<TErrorShape>,
 ) => createBuilder<"web", TErrorShape>(opts);
+
+export const createServerHandler = <TRouter extends FileRouter>(
+  opts: RouterWithConfig<TRouter>,
+) => {
+  const requestHandler = buildRequestHandler<TRouter>(opts);
+
+  const POST = async (request: Request | { request: Request }) => {
+    const req = request instanceof Request ? request : request.request;
+    const response = await requestHandler({ req });
+
+    if (response instanceof UploadThingError) {
+      const errorFormatter =
+        opts.router[Object.keys(opts.router)[0]]?._def.errorFormatter ??
+        defaultErrorFormatter;
+
+      return new Response(JSON.stringify(errorFormatter(response)), {
+        status: getStatusCodeFromError(response),
+        headers: {
+          "x-uploadthing-version": UPLOADTHING_VERSION,
+        },
+      });
+    }
+    if (response.status !== 200) {
+      // We messed up - this should never happen
+      return new Response("An unknown error occured", {
+        status: 500,
+        headers: {
+          "x-uploadthing-version": UPLOADTHING_VERSION,
+        },
+      });
+    }
+
+    return new Response(JSON.stringify(response.body), {
+      status: response.status,
+      headers: {
+        "x-uploadthing-version": UPLOADTHING_VERSION,
+      },
+    });
+  };
+
+  const getBuildPerms = buildPermissionsInfoHandler<TRouter>(opts);
+
+  const GET = (request: Request | { request: Request }) => {
+    const _req = request instanceof Request ? request : request.request;
+
+    return new Response(JSON.stringify(getBuildPerms()), {
+      status: 200,
+      headers: {
+        "x-uploadthing-version": UPLOADTHING_VERSION,
+      },
+    });
+  };
+
+  return { GET, POST };
+};
 
 export const extractRouterConfig = (router: FileRouter) =>
   buildPermissionsInfoHandler({ router })();
